noir_ast_fuzzer_fuzz

Module targets

Modules

acir_vs_brillig

Compare the execution of random ASTs between the normal execution vs when everything is forced to be Brillig.

comptime_vs_brillig_direct

Compare the execution of random ASTs between the comptime execution (after converting the AST to Noir and running it through the comptime interpreter) vs when everything is forced to be Brillig. We choose Brillig here because it mostly matches comptime feature set (e.g. loop, while, break and continue are possible) This variant accesses the interpreter directly instead of going through nargo, which speeds up execution but also currently has some issues (inability to use prints among others).

comptime_vs_brillig_nargo

Compare the execution of random ASTs between the comptime execution (after converting the AST to Noir and running it through the comptime interpreter) vs when everything is forced to be Brillig. We choose Brillig here because it mostly matches comptime feature set (e.g. loop, while, break and continue are possible) This variant lets nargo parse the resulting source code which is slow but at the moment is more feature complete than using the interpreter directly.

min_vs_full

Compare the execution of random ASTs between the initial SSA (or as close as we can stay to the initial state) and the fully optimized version.

orig_vs_morph

Perform random metamorphic mutations on the AST and check that the execution result does not change.

pass_vs_prev

Compare the execution of an SSA pass to the one preceding it using the SSA interpreter.

Functions

default_config 🔒

Create a default configuration instance, with some common flags randomized.