noir_ast_fuzzer_fuzz::targets
Module comptime_vs_brillig_nargo

Compare the execution of random ASTs between the comptime execution (after converting the AST to Noir and running it through the comptime interpreter) vs when everything is forced to be Brillig. We choose Brillig here because it mostly matches comptime feature set (e.g. loop, while, break and continue are possible) This variant lets nargo parse the resulting source code which is slow but at the moment is more feature complete than using the interpreter directly.

Functions

fuzz