noir_ast_fuzzer_fuzz::targets

Module orig_vs_morph

Perform random metamorphic mutations on the AST and check that the execution result does not change.

Modules

helpers 🔒

rules 🔒

Metamorphic transformation rules.

Structs

MorphContext 🔒

Recursively apply rules while keeping a tally on how many we have done.

VariableContext 🔒

Context necessary to generate new local IDs during rewrites.

Functions

estimate_applicable_rules 🔒

Provide a rough estimate for how many rules can be applied.

fuzz

is_special_call 🔒

Check if we are calling an oracle or builtin function.

rewrite_function 🔒

rewrite_program 🔒