noir_ast_fuzzer_fuzz::targets::orig_vs_morph
Module rules

Metamorphic transformation rules.

Structs

Context

Rule

Metamorphic transformation rule.

Functions

any_inevitable

Transform any expression into an if-then-else with the itself repeated in the then and else branch:

bool_or_self

Transform boolean value x into x | x.

bool_rule_matches 🔒

Common match condition for boolean rules.

bool_xor_rand

Transform boolean value x into rnd ^ x ^ rnd.

bool_xor_self

Transform boolean value x into x ^ x ^ x.

collect

Construct all rules that we can apply on a program.

int_break_up

Break an integer literal a into b + c.

num_add_zero

Transform any numeric value x into x+0

num_commute

Transform commutative arithmetic operations:

num_div_one

Transform any numeric value x into x/1

num_mul_one

Transform any numeric value x into x*1

num_op 🔒

Transform any numeric value x into x <op> <rhs>

num_rule_matches 🔒

Common condition for numeric rules

num_sub_zero

Transform any numeric value x into x-0

Type Aliases

MatchFn 🔒

Check if the rule can be applied on an expression.

RewriteFn 🔒

Apply the rule on an expression, mutating/replacing it in-place.