Crate noir_ast_fuzzer_fuzz

Modules

targets

Functions

bool_from_env 🔒

compare_results_compiled

Compare the execution result and print the inputs if the result is a failure.

compare_results_comptime

Compare the execution result for comptime fuzzing and print the inputs if the result is a failure.

compare_results_interpreted

Compare the execution result and print the inputs if the result is a failure.

compile_into_circuit_or_die

Compile a monomorphized Program into circuit or panic.

compile_into_circuit_with_ssa_passes_or_die

Compile a monomorphized Program into circuit using the given SSA passes or panic.

default_ssa_options

show_ssa 🔒

Show all SSA passes during compilation.