noir_ast_fuzzer_fuzz::targets::orig_vs_morph
Module helpers

Statics

TYPES 🔒

Types we can consider using in this context.

Functions

gen_binary 🔒

Generate an arbitrary binary expression, returning a specific type.

gen_expr 🔒

Generate an arbitrary pure (free of side effects) expression, returning a specific type.

gen_unary 🔒

Generate an arbitrary unary expression, returning a specific type.

has_side_effect 🔒

Check if an expression can have a side effect, in which case duplicating or reordering it could change the behavior of the program. This doesn’t concern about failures, just observable changes the state of the program.

reassign_ids 🔒

Assign new IDs to variables and identifiers created in the expression.